requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
gpiozero==2.0.1
lgpio==0.2.2.0
spidev==3.8
pillow==12.0.0
pigpio==1.78
waveshare-epaper==1.3.0
